+++ web/src/components/ui/index.ts
@@ -731,3 +731,4 @@ export * from './chart-line-macd-zero-cross-sig';
 export * from './chart-line-hma-cross-sig';
 export * from './chart-line-ema-cross-sig';
 export * from './chart-line-tema-cross-sig';
+export * from './chart-line-dema-cross-sig';